diktat fixes
