+++ save-backend/src/main/kotlin/org/cqfn/save/backend/controllers/AwesomeBenchmarksController.kt
@@ -1,16 +1,29 @@
+/**
+ * Controller for processing awesome-benchmarks:
+ * 1) to put benchamrks to DB
+ * 2) to get benchmarks for the Frontend
+ */
+
 package org.cqfn.save.backend.controllers
 
-import org.cqfn.save.entities.benchmarks.BenchmarkEntity
 import org.cqfn.save.backend.service.AwesomeBenchmarksService
+import org.cqfn.save.entities.benchmarks.BenchmarkEntity
+
 import org.slf4j.LoggerFactory
 import org.springframework.web.bind.annotation.*
 
 @RestController
+/**
+ * Controller for processing awesome-benchmarks
+ */
 class AwesomeBenchmarksController(
-        private val awesomeBenchmarksService: AwesomeBenchmarksService,
+    private val awesomeBenchmarksService: AwesomeBenchmarksService,
 ) {
     private val log = LoggerFactory.getLogger(AwesomeBenchmarksController::class.java)
 
+    /**
+     * @param benchmarks
+     */
     @PostMapping("/internal/upload/awesome-benchmarks")
     fun uploadAwesomeBenchmarks(@RequestBody(required = true) benchmarks: List<BenchmarkEntity>) {
         log.info("Received a request to save awesome-benchmarks to the db")
@@ -18,7 +31,10 @@ class AwesomeBenchmarksController(
         log.info("Saved requested awesome-benchmarks to the db")
     }
 
+    /**
+     * @return all benchmarks from backend to frontend
+     */
     @GetMapping("/api/awesome-benchmarks")
     fun getAllAwesomeBenchmarks() =
-        awesomeBenchmarksService.getAllBenchmarks()
+            awesomeBenchmarksService.getAllBenchmarks()
 }
