+++ save-cloud-common/src/commonMain/kotlin/org/cqfn/save/entities/benchmarks/BenchmarkEntity.kt
@@ -1,41 +1,71 @@
+/**
+ * Common class for representing the entity used in awesome-benchamrks
+ */
+
 package org.cqfn.save.entities.benchmarks
 
 import kotlinx.serialization.Serializable
 
 @Serializable
+/**
+ * Enum that represents categories that could be used for grouping benchmarks (they should be selected for each benchmark)
+ */
 enum class BenchmarkCategoryEnum {
-    STATIC_ANALYSIS,
+    AI,
+    AUDIT,
     CODING_STANDARD,
     PERFORMANCE,
-    AUDIT,
-    AI,
+    STATIC_ANALYSIS,
+    ;
 }
 
+/**
+ * @property general
+ * @property info
+ * @property links
+ */
 @Serializable
 data class BenchmarkEntity(
-        val general: General,
-        val info: Info,
-        val links: Links,
+    val general: General,
+    val info: Info,
+    val links: Links,
 )
 
+/**
+ * @property name
+ * @property category
+ * @property tags
+ */
 @Serializable
 data class General(
-        val name: String,
-        val category: BenchmarkCategoryEnum,
-        val tags: ArrayList<String>,
+    val name: String,
+    val category: BenchmarkCategoryEnum,
+    val tags: ArrayList<String>,
 )
 
+/**
+ * @property language
+ * @property license
+ * @property scenarios_num
+ * @property description
+ */
 @Serializable
 data class Info(
-        val language: String,
-        val license: String,
-        val scenarios_num: Long,
-        val description: String,
+    val language: String,
+    val license: String,
+    @Suppress("VARIABLE_NAME_INCORRECT_FORMAT", "ConstructorParameterNaming")
+    val scenarios_num: Long,
+    val description: String,
 )
 
+/**
+ * @property homepage
+ * @property sources
+ * @property documentation
+ */
 @Serializable
 data class Links(
-        val homepage: String,
-        val sources: String,
-        val documentation: String,
+    val homepage: String,
+    val sources: String,
+    val documentation: String,
 )
