+++ save-frontend/src/main/kotlin/org/cqfn/save/frontend/externals/fontawesome/Icons.kt
@@ -35,4 +35,3 @@ external val faCheckCircle: dynamic
 external val faPlus: dynamic
 external val faArrowRight: dynamic
 external val faLightbulb: dynamic
-
