+++ save-preprocessor/src/main/kotlin/org/cqfn/save/preprocessor/utils/GitUtil.kt
@@ -1,3 +1,7 @@
+/**
+ * Git utilities that are used in preprocessor for download/clone/update git repo
+ */
+
 package org.cqfn.save.preprocessor.utils
 
 import org.cqfn.save.entities.GitDto
@@ -6,6 +10,11 @@ import org.eclipse.jgit.transport.CredentialsProvider
 import org.eclipse.jgit.transport.UsernamePasswordCredentialsProvider
 import java.io.File
 
+/**
+ * @param gitDto
+ * @param tmpDir
+ * @return jGit git entity
+ */
 fun cloneFromGit(gitDto: GitDto, tmpDir: File): Git? {
     val userCredentials = if (gitDto.username != null && gitDto.password != null) {
         UsernamePasswordCredentialsProvider(gitDto.username, gitDto.password)
@@ -13,8 +22,8 @@ fun cloneFromGit(gitDto: GitDto, tmpDir: File): Git? {
         CredentialsProvider.getDefault()
     }
     return Git.cloneRepository()
-            .setURI(gitDto.url)
-            .setCredentialsProvider(userCredentials)
-            .setDirectory(tmpDir)
-            .call()
+        .setURI(gitDto.url)
+        .setCredentialsProvider(userCredentials)
+        .setDirectory(tmpDir)
+        .call()
 }
